requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
openai>=1.52.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2